Remove DZil plugins I don't _really_ need.
This repo is essentially _just_ for the YAML files, I'm just cheating
and using DZil and a `.pm` file to help generate the `README.md` and
`LICENSE` files.

So, don't need the other DZil plugins.

I do, however, want to make sure that if I'm going to roll a release,
that I indicate the changes, so add the `EnsureChangesHasContent`
plugin.