Add a step to dump the Perl version being used.
Just because we've asked for "5.10" or "5.30" doesn't give us all the
info we may want to help diagnose a problem.  So, dump the full output
of `perl -V` as a build step, so we have that handy in the event that we
need to know exactly which minor version of Perl is installed and in
use.